+++ tests/test_b32.py
@@ -316,8 +316,8 @@ def test_p2wsh() -> None:
 
 def test_p2tr() -> None:
     key = 1
-    pubkey = output_pubkey(key)[0]
+    pub_key = output_pubkey(key)[0]
     addr = b32.p2tr(key)
     _, wit_prg, _ = b32.witness_from_address(addr)
 
-    assert wit_prg == pubkey
+    assert wit_prg == pub_key